+++ HomeMyDay.Web.Api/Api/PagesController.cs
@@ -25,7 +25,7 @@ namespace HomeMyDay.Web.Api.Controllers
 
 		// GET api/values
 		[HttpGet("{id}")]
-		public Page Get(int id)
+		public Page Get(long id)
         {
 			return pageManager.GetPage(id);
         }
@@ -62,7 +62,7 @@ namespace HomeMyDay.Web.Api.Controllers
 
         // PUT api/values/5
         [HttpPut("{id}")]
-        public IActionResult Put(int id, [FromBody]Page page)
+        public IActionResult Put(long id, [FromBody]Page page)
         {
 			if (!ModelState.IsValid)
 			{
@@ -93,7 +93,7 @@ namespace HomeMyDay.Web.Api.Controllers
 
 		// DELETE api/values/5
 		[HttpDelete("{id}")]
-        public IActionResult Delete(int id)
+        public IActionResult Delete(long id)
         {
 			pageManager.DeletePage(id);
